refactor(app): unify auth state cleanup via jotai store
